Add clang-tidy to CI and fix what it found
Phase 2 track 4. .clang-tidy selects a narrow, bug-focused check set
rather than the broad modernize-*/readability-* families, which would
flag thousands of lines of 2005-era style with no correctness benefit -
and a gate nobody can keep green gets disabled. Every exclusion is
justified in the file against the actual findings, not guessed.

Two real issues came out of it:

ParserJuly2004.cpp appended istream::get() results, which are int,
straight onto std::string. That works only because each call site has
already checked the value is not EOF - the same implicit int-to-char
narrowing whose absence caused the infinite loop fixed in 5.1.0, in the
same file. Now cast explicitly at all 13 sites.

Base64Encoder.cpp computed the output buffer size in int arithmetic
before widening it for the allocation, so a large enough input would
overflow and under-allocate. Widened before the multiply.

Worth recording: clang-tidy's own suggested fix for the first issue was
to wrap the value in std::to_string(), which renders the character code
as decimal text - a space would have become "32". The test suite caught
it immediately (roundtrip_fidelity, contract and enumeration all
failed), which is a good argument for not applying --fix blindly.

CI gains a format job and a tidy job. .clang-tidy sets
WarningsAsErrors, so any new finding fails the build.

+++ src/Base64Encoder.cpp
@@ -92,7 +92,11 @@ char *Base64Encoder::encode(const char *data, int datalength, int &outlength)
 		return (char *) 0;
 	};
 
-	char *retval = new char[datalength * 2];
+	// Widen before multiplying: datalength is an int, so the product
+	// overflowed in int arithmetic before being widened for new[],
+	// which would under-allocate for very large inputs.
+	//
+	char *retval = new char[(size_t) datalength * 2];
 	const char *crlf = "\n";
 	int crlflength = strlen(crlf);
 	int maxlinelength = 76;
